requirements.txt
canonicalwebteam.http==0.1.4
requests-cache==0.4.13
pyyaml==3.13
yamlordereddictloader==0.4.0
Flask==0.12.2
beautifulsoup4==4.6.3
prometheus-flask-exporter==0.2.0
statsd<3.3.0
gunicorn[gevent]
talisker==0.9.14
